+++ babel.config.js
@@ -1,6 +1,10 @@
 module.exports = {
-  presets: [
-    '@vue/cli-plugin-babel/preset'
+  "presets": [
+    ["@vue/cli-plugin-babel/preset", 
+      {
+      "useBuiltIns": false
+      }
+    ]
   ],
   plugins: ['@babel/transform-runtime', 'babel-plugin-transform-import-meta'],
 }